Remove old history
