+++ web/src/types/preview.ts
@@ -1,0 +1,7 @@
+type Preview = {
+    camera: string;
+    src: string;
+    type: string;
+    start: number;
+    end: number;
+  };
